Ya muestra la persona

+++ actividad1/src/main/java/com/mycompany/actividad1/dao/PersonaDAO.java
@@ -1,11 +1,11 @@
 package com.mycompany.actividad1.dao;
 
-
 import com.mycompany.actividad1.model.Persona;
 import com.mycompany.actividad1.Database;
 import java.sql.*;
 import java.util.ArrayList;
 import java.util.List;
+
 
 public class PersonaDAO {
 
@@ -13,7 +13,6 @@ public class PersonaDAO {
         String sql = "INSERT INTO persona (nombres, apellidos, email) VALUES (?, ?, ?)";
         try (Connection conn = Database.getConnection();
              PreparedStatement stmt = conn.prepareStatement(sql)) {
-
             stmt.setString(1, persona.getNombres());
             stmt.setString(2, persona.getApellidos());
             stmt.setString(3, persona.getEmail());
@@ -31,7 +30,7 @@ public class PersonaDAO {
 
             while (rs.next()) {
                 Persona persona = new Persona(
-                        rs.getLong("id"),
+                        rs.getDouble("id"),          
                         rs.getString("nombres"),
                         rs.getString("apellidos"),
                         rs.getString("email")
@@ -46,7 +45,6 @@ public class PersonaDAO {
         String sql = "DELETE FROM persona WHERE id = ?";
         try (Connection conn = Database.getConnection();
              PreparedStatement stmt = conn.prepareStatement(sql)) {
-
             stmt.setLong(1, id);
             stmt.executeUpdate();
         }
